Refactor forMonth() as forMonthContaining().
Instead of forMonth() taking a date, it could be refactored to take a year and month (integers); to keep the ability to provide a date within the month, i have created a method called forMonthContaining().

The forMonthContaining() method could also accept a Carbon date instance as well as the previous string and format.